Half ton MG Fix - Read description note!
You had two mg boxes, one ending with "_half" and the other ending with "_Half" Ids are case sensitive so there are two otherwise identical ammo boxes in use. I have standardized on _half, lowercase. The upper case one has been marked with a warning encouraging players to sell it.  In the meantime it's left as a fallback so it doesn't break the saves of anyone who has it in their inventory.